break view into smaller components
